Issue 11 - Coding The EMAC Rule/kq-datahub/requirements.txt
pandas
psycopg2
python-dotenv
matplotlib